[Test] Exercise LTX vocoder CUDA autocast
